requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.36.0
google.generativeai==0.8.3
